Modify get users endpoint to return users in same department
